Try to backport anyway, regardless of conflict label
Prior to this commit, the code would check for the conflict label and
just stop processing. Instead, this commit changes it so that it will
try and backport anyway. This helps in cases where someone marks a
different PR for backport, which frees up the conflict PR, so they don't
need to actually remove the conflict. Checking for a conflict is not
expensive, and if we find we are still in conflict, then we can check if
the issue is already labeled or not.