Fix YAML syntax for multiple merge keys
Multiple instances of `<<:` for merge keys was removed in the YAML 1.2
spec, but many parsers still accept it.  However, vscode only supports
the newer YAML spec, and so these lines show up as errors.  Instead, we
can use an array of references within the merge key.

See also https://stackoverflow.com/a/64228697/2355944